LPD-78127 Move FDSConnection logic into main liferay-portal project leaving only types here. Adding a bridge to let connect with frontend-data-set-web in runtime
